% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_appview.R
\name{appview}
\alias{appview}
\alias{has_bluesky_appview}
\alias{get_bluesky_appview}
\alias{bs_get_appview}
\alias{bs_has_appview}
\title{Check or Get Bluesky AppView}
\usage{
has_bluesky_appview()

get_bluesky_appview()

bs_get_appview()

bs_has_appview()
}
\value{
logical if \code{has}, AppView URL if \code{get}
}
\description{
Check or Get Bluesky AppView
}
\examples{
has_bluesky_appview()
}
\concept{auth}
