% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/created_at.R
\name{bs_created_at}
\alias{bs_created_at}
\title{Get current time in Bluesky format}
\usage{
bs_created_at()
}
\value{
a length 1 character vector
}
\description{
Get current time in Bluesky format
}
\section{Function introduced}{

\code{v0.1.0} (2023-11-25)
}

\examples{
bs_created_at()
}
\concept{helper}
