% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_uri.R
\name{bs_extract_record_key}
\alias{bs_extract_record_key}
\title{Extract Record Key from a link}
\usage{
bs_extract_record_key(url)
}
\arguments{
\item{url}{Character, length 1. URL for record to get.}
}
\value{
character vector of record keys
}
\description{
Extract Record Key from a link
}
\examples{
bs_extract_record_key('https://bsky.app/profile/chriskenny.bsky.social/post/3lc5d6zspys2c')
}
