% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_feed_generator.R
\name{bs_get_feed_generator}
\alias{bs_get_feed_generator}
\title{Get specific information about one feed generator}
\usage{
bs_get_feed_generator(
  feed,
  user = get_bluesky_user(),
  pass = get_bluesky_pass(),
  auth = bs_auth(user, pass),
  clean = TRUE
)
}
\arguments{
\item{feed}{Character, length 1. Feed to get.}

\item{user}{Character. User name to log in with. Defaults to \code{get_bluesky_user()}.}

\item{pass}{Character. App password to log in with. Defaults to \code{get_bluesky_pass()}.}

\item{auth}{Authentication information. Defaults to \code{bs_auth(user, pass)}.}

\item{clean}{Logical. Should output be cleaned into a \code{tibble}? Default: \code{TRUE}.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} of feeds
}
\description{
Get specific information about one feed generator
}
\section{Lexicon references}{

\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/feed/getFeedGenerator.json}{feed/getFeedGenerator.json (2023-10-01)}
}

\section{Function introduced}{

\code{v0.0.1} (2023-10-01)
}

\examples{
\dontshow{if (has_bluesky_pass() && has_bluesky_user()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bs_get_feed_generator('at://did:plc:z72i7hdynmk6r22z27h6tvur/app.bsky.feed.generator/bsky-team')
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=bs_get_feed_generators]{bs_get_feed_generators()}} for less detailed information about multiple feed generators.
}
\concept{feed}
