% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actor_preferences.R
\name{bs_get_preferences}
\alias{bs_get_preferences}
\title{Get (Self) Preferences}
\usage{
bs_get_preferences(user = get_bluesky_user(), pass = get_bluesky_pass())
}
\arguments{
\item{user}{Character. User name to log in with. Defaults to \code{get_bluesky_user()}.}

\item{pass}{Character. User name to log in with. Defaults to \code{get_bluesky_user()}.}
}
\value{
a tibble of preferences
}
\description{
Get (Self) Preferences
}
\examples{
\dontshow{if (has_bluesky_pass() && has_bluesky_user()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bs_get_preferences()
\dontshow{\}) # examplesIf}
}
\concept{actor}
