% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actor_suggestions.R
\name{bs_get_suggestions}
\alias{bs_get_suggestions}
\title{Get a list of actors suggested for following}
\usage{
bs_get_suggestions(user = get_bluesky_user(), pass = get_bluesky_pass())
}
\arguments{
\item{user}{Character. User name to log in with. Defaults to \code{get_bluesky_user()}.}

\item{pass}{Character. User name to log in with. Defaults to \code{get_bluesky_user()}.}
}
\value{
a tibble of suggested accounts to follow
}
\description{
Get a list of actors suggested for following
}
\examples{
\dontshow{if (has_bluesky_pass() & has_bluesky_user()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bs_get_suggestions
\dontshow{\}) # examplesIf}
}
\concept{actor}
