% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_get_upload_limits.R
\name{bs_get_video_upload_limits}
\alias{bs_get_video_upload_limits}
\title{Get Video Upload Limits}
\usage{
bs_get_video_upload_limits(
  user = get_bluesky_user(),
  pass = get_bluesky_pass(),
  auth = bs_auth(user, pass),
  clean = TRUE
)
}
\arguments{
\item{user}{Character. User name to log in with. Defaults to \code{get_bluesky_user()}.}

\item{pass}{Character. App password to log in with. Defaults to \code{get_bluesky_pass()}.}

\item{auth}{Authentication information. Defaults to \code{bs_auth(user, pass)}.}

\item{clean}{Logical. Should output be cleaned into a \code{tibble}? Default: \code{TRUE}.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} of video upload allowances, or a list if \code{clean = FALSE}
}
\description{
Get Video Upload Limits
}
\section{Lexicon references}{

\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/video/getUploadLimits.json}{video/getUploadLimits.json (2024-11-23)}
}

\section{Function introduced}{

\code{v0.5.0} (2026-06-01)
}

\examples{
\dontshow{if (has_bluesky_pass() && has_bluesky_user()) withAutoprint(\{ # examplesIf}
bs_get_video_upload_limits()
\dontshow{\}) # examplesIf}
}
\concept{video}
