% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convo_send_message_batch.R
\name{bs_send_message_batch}
\alias{bs_send_message_batch}
\title{Send multiple messages across different conversations}
\usage{
bs_send_message_batch(
  convo_id,
  text,
  user = get_bluesky_user(),
  pass = get_bluesky_pass(),
  auth = bs_auth(user, pass),
  clean = TRUE
)
}
\arguments{
\item{convo_id}{Character, length 1. ID of the conversation to get.}

\item{text}{Character vector of message texts.}

\item{user}{Character. User name to log in with. Defaults to \code{get_bluesky_user()}.}

\item{pass}{Character. App password to log in with. Defaults to \code{get_bluesky_pass()}.}

\item{auth}{Authentication information. Defaults to \code{bs_auth(user, pass)}.}

\item{clean}{Logical. Should output be cleaned into a \code{tibble}? Default: \code{TRUE}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} or a \code{list} if \code{clean = FALSE}.
}
\description{
Send multiple messages across different conversations
}
\section{Lexicon references}{

\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/chat/bsky/convo/sendMessageBatch.json}{chat.bsky.convo.sendMessageBatch.json (2025-05-16)}
}

\section{Function introduced}{

\code{v0.3.1} (2025-05-16)
}

\examples{
\dontshow{if (has_bluesky_pass() && has_bluesky_user()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bs_send_message_batch(
  convo_id = c('3ku7w6h4vog2d', '3lpidxucy2g27'),
  text = c('Hello', 'Hi there')
)
\dontshow{\}) # examplesIf}
}
\concept{chat}
