% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_upload_blob.R
\name{bs_upload_blob}
\alias{bs_upload_blob}
\title{Upload a blob to a repo}
\usage{
bs_upload_blob(
  blob,
  user = get_bluesky_user(),
  pass = get_bluesky_pass(),
  auth = bs_auth(user, pass)
)
}
\arguments{
\item{blob}{TODO}

\item{user}{Character. User name to log in with. Defaults to \code{get_bluesky_user()}.}

\item{pass}{Character. App password to log in with. Defaults to \code{get_bluesky_pass()}.}

\item{auth}{Authentication information. Defaults to \code{bs_auth(user, pass)}.}
}
\value{
a tibble with a single column and row for the count
}
\description{
Upload a blob to a repo
}
\section{Lexicon references}{

\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/com/atproto/repo/uploadBlob.json}{repo/uploadBlob.json (2023-10-02)}
}

\section{Function introduced}{

\code{v0.0.1} (2023-10-02)
}

\examples{
\dontshow{if (has_bluesky_pass() & has_bluesky_user()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bs_upload_blob()
\dontshow{\}) # examplesIf}
}
\concept{repo}
