% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_upload_video.R
\name{bs_upload_video}
\alias{bs_upload_video}
\title{Upload a video to be processed and stored}
\usage{
bs_upload_video(
  video,
  user = get_bluesky_user(),
  pass = get_bluesky_pass(),
  auth = bs_auth(user, pass)
)
}
\arguments{
\item{video}{Character, videos to upload to a repo.}

\item{user}{Character. User name to log in with. Defaults to \code{get_bluesky_user()}.}

\item{pass}{Character. App password to log in with. Defaults to \code{get_bluesky_pass()}.}

\item{auth}{Authentication information. Defaults to \code{bs_auth(user, pass)}.}
}
\value{
a blob reference list, compatible with \code{bs_post()}
}
\description{
Upload a video to be processed and stored
}
\section{Lexicon references}{

\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/video/uploadVideo.json}{video/uploadVideo.json (2024-11-23)}
\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/video/getJobStatus.json}{video/getJobStatus.json (2024-11-23)}
}

\section{Function introduced}{

\code{v0.5.0} (2026-06-01)
}

\examples{
\dontshow{if (has_bluesky_pass() & has_bluesky_user()) withAutoprint(\{ # examplesIf}
vid <- fs::path_package('bskyr', 'man/figures/pkgs.mp4')
bs_upload_video(vid)
\dontshow{\}) # examplesIf}
}
\concept{repo}
