% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_uri.R
\name{bs_uri_to_url}
\alias{bs_uri_to_url}
\title{Convert Universal Resource Identifiers to Hypertext Transfer Protocol Secure URLs}
\usage{
bs_uri_to_url(uri)
}
\arguments{
\item{uri}{Character, length 1. URI for post to get.}
}
\value{
character vector of HTTPS URLs
}
\description{
Convert Universal Resource Identifiers to Hypertext Transfer Protocol Secure URLs
}
\examples{
bs_uri_to_url('at://did:plc:ic6zqvuw5ulmfpjiwnhsr2ns/app.bsky.feed.post/3k7qmjev5lr2s')
}
