% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_pds.R
\name{pds}
\alias{pds}
\alias{has_bluesky_pds}
\alias{get_bluesky_pds}
\alias{bs_get_pds}
\alias{bs_has_pds}
\title{Check or Get Bluesky PDS}
\usage{
has_bluesky_pds()

get_bluesky_pds()

bs_get_pds()

bs_has_pds()
}
\value{
logical if \code{has}, PDS URL if \code{get}
}
\description{
Check or Get Bluesky PDS
}
\examples{
has_bluesky_pds()
}
\concept{auth}
