% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_appview.R
\name{set_bluesky_appview}
\alias{set_bluesky_appview}
\alias{bs_set_appview}
\title{Adds Bluesky AppView to .Renviron.}
\usage{
set_bluesky_appview(appview, overwrite = FALSE, install = FALSE, r_env = NULL)

bs_set_appview(appview, overwrite = FALSE, install = FALSE, r_env = NULL)
}
\arguments{
\item{appview}{Character. Base URL of the AppView server to add.}

\item{overwrite}{Defaults to FALSE. Boolean. Should existing \code{BLUESKY_APP_APPVIEW} in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Boolean. Should this be added '~/.Renviron' file?}

\item{r_env}{Path to install to if \code{install} is \code{TRUE}.}
}
\value{
appview, invisibly
}
\description{
Adds Bluesky AppView to .Renviron.
}
\examples{
example_env <- tempfile(fileext = '.Renviron')
set_bluesky_appview('https://bsky.social', r_env = example_env)
# r_env should likely be: file.path(Sys.getenv('HOME'), '.Renviron')
}
\concept{auth}
