% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_pass.R
\name{set_bluesky_pass}
\alias{set_bluesky_pass}
\alias{bs_set_pass}
\title{Add Entry to Renviron}
\usage{
set_bluesky_pass(pass, overwrite = FALSE, install = FALSE)

bs_set_pass(pass, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{pass}{Character. App Password to add to add.}

\item{overwrite}{Defaults to FALSE. Boolean. Should existing \code{BLUESKY_APP_PASS} in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Boolean. Should this be added '~/.Renviron' file?}
}
\value{
pass, invisibly
}
\description{
Adds Bluesky App Password to .Renviron.
}
\examples{
\dontrun{
set_bluesky_pass('1234')
}

}
\concept{auth}
