% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_pds.R
\name{set_bluesky_pds}
\alias{set_bluesky_pds}
\alias{bs_set_pds}
\title{Adds Bluesky PDS to .Renviron.}
\usage{
set_bluesky_pds(pds, overwrite = FALSE, install = FALSE, r_env = NULL)

bs_set_pds(pds, overwrite = FALSE, install = FALSE, r_env = NULL)
}
\arguments{
\item{pds}{Character. Base URL of the Personal Data Server to add.}

\item{overwrite}{Defaults to FALSE. Boolean. Should existing \code{BLUESKY_APP_PDS} in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Boolean. Should this be added '~/.Renviron' file?}

\item{r_env}{Path to install to if \code{install} is \code{TRUE}.}
}
\value{
pds, invisibly
}
\description{
Adds Bluesky PDS to .Renviron.
}
\examples{
example_env <- tempfile(fileext = '.Renviron')
set_bluesky_pds('https://bsky.social', r_env = example_env)
# r_env should likely be: file.path(Sys.getenv('HOME'), '.Renviron')
}
\concept{auth}
