% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\alias{has_bluesky_user}
\alias{get_bluesky_user}
\title{Check or Get Bluesky User}
\usage{
has_bluesky_user()

get_bluesky_user()
}
\value{
logical if \code{has}, user if \code{get}
}
\description{
Check or Get Bluesky User
}
\examples{
has_bluesky_user()
}
