% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bs_auth}
\alias{bs_auth}
\title{Authenticate a user}
\usage{
bs_auth(user, pass)
}
\arguments{
\item{user}{Character. User name to log in with.}

\item{pass}{Character. App password to log in with.}
}
\value{
a list of authentication information
}
\description{
Authenticate a user
}
\section{Lexicon references}{

\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/com/atproto/server/createSession.json}{server/createSession.json (2023-09-30)}
}

\section{Function introduced}{

\code{v0.0.1} (2023-09-30)
}

\examples{
\dontshow{if (has_bluesky_pass() && has_bluesky_user()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bs_auth(user = get_bluesky_user(), pass = get_bluesky_pass())
\dontshow{\}) # examplesIf}
}
\concept{auth}
