% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_feed_generator.R, R/feed_feed_generators.R
\name{bs_get_feed_generator}
\alias{bs_get_feed_generator}
\title{Get specific information about one feed generator}
\usage{
bs_get_feed_generator(
  feeds,
  user = get_bluesky_user(),
  pass = get_bluesky_pass(),
  auth = bs_auth(user, pass)
)

bs_get_feed_generator(
  feeds,
  user = get_bluesky_user(),
  pass = get_bluesky_pass(),
  auth = bs_auth(user, pass)
)
}
\arguments{
\item{feeds}{Character, length 1. Vector of feeds to get.}

\item{user}{Character. User name to log in with. Defaults to \code{get_bluesky_user()}.}

\item{pass}{Character. App password to log in with. Defaults to \code{get_bluesky_pass()}.}

\item{auth}{Authentication information. Defaults to \code{bs_auth(user, pass)}.}

\item{feed}{Character, length 1. Feed to get.}
}
\value{
a tibble of feeds

a tibble of feeds
}
\description{
Get specific information about one feed generator

Get information about a list of feed generators
}
\section{Lexicon references}{

\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/feed/getFeedGenerator.json}{feed/getFeedGenerator.json (2023-10-01)}


\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/feed/getFeedGenerators.json}{feed/getFeedGenerators.json (2023-10-01)}
}

\section{Function introduced}{

\code{v0.0.1} (2023-10-01)


\code{v0.0.1} (2023-10-01)
}

\examples{
\dontshow{if (has_bluesky_pass() && has_bluesky_user()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bs_get_feed_generator('at://did:plc:z72i7hdynmk6r22z27h6tvur/app.bsky.feed.generator/bsky-team')
\dontshow{\}) # examplesIf}
\dontshow{if (has_bluesky_pass() && has_bluesky_user()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bs_get_feed_generator('at://did:plc:z72i7hdynmk6r22z27h6tvur/app.bsky.feed.generator/bsky-team')
\dontshow{\}) # examplesIf}
}
\concept{feed}
