% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_search.R
\name{bs_search_posts}
\alias{bs_search_posts}
\title{Find posts matching search criteria}
\usage{
bs_search_posts(
  query,
  cursor = NULL,
  limit = NULL,
  user = get_bluesky_user(),
  pass = get_bluesky_pass(),
  auth = bs_auth(user, pass),
  clean = TRUE
)
}
\arguments{
\item{query}{character. search query, Lucene query syntax is recommended.}

\item{cursor}{Character, length 1. A cursor property from a prior response. Default: \code{NULL}.}

\item{limit}{Integer. Number of records to request. If over \code{100}, multiple requests are made.}

\item{user}{Character. User name to log in with. Defaults to \code{get_bluesky_user()}.}

\item{pass}{Character. App password to log in with. Defaults to \code{get_bluesky_pass()}.}

\item{auth}{Authentication information. Defaults to \code{bs_auth(user, pass)}.}

\item{clean}{Logical. Should output be cleaned into a \code{tibble}? Default: \code{TRUE}.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} of suggested accounts to follow
}
\description{
Find posts matching search criteria
}
\section{Lexicon references}{

\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/feed/searchPosts.json}{feed/searchPosts.json (2023-12-13)}
}

\section{Function introduced}{

\code{v0.1.1} (2023-12-13)
}

\examples{
\dontshow{if (has_bluesky_pass() & has_bluesky_user()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bs_search_posts('redistricting')
\dontshow{\}) # examplesIf}
}
\concept{feed}
