% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_user.R
\name{set_bluesky_user}
\alias{set_bluesky_user}
\alias{bs_set_user}
\title{Add Entry to Renviron}
\usage{
set_bluesky_user(user, overwrite = FALSE, install = FALSE)

bs_set_user(user, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{user}{Character. User to add to add.}

\item{overwrite}{Defaults to FALSE. Boolean. Should existing \code{BLUESKY_APP_USER} in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Boolean. Should this be added '~/.Renviron' file?}
}
\value{
user, invisibly
}
\description{
Adds Bluesky User to .Renviron.
}
\examples{
\dontrun{
set_bluesky_user('1234')
}

}
